+++ app/components/UI/TransactionElement/index.js
@@ -4,22 +4,9 @@ import { TouchableHighlight, StyleSheet, Text, View, Image } from 'react-native'
 import { colors, fontStyles } from '../../../styles/common';
 import { strings } from '../../../../locales/i18n';
 import { toLocaleDateTime } from '../../../util/date';
-import {
-	renderFromWei,
-	weiToFiat,
-	hexToBN,
-	toBN,
-	isBN,
-	renderToGwei,
-	balanceToFiat,
-	renderFromTokenMinimalUnit,
-	fromTokenMinimalUnit,
-	balanceToFiatNumber,
-	addCurrencySymbol,
-	weiToFiatNumber
-} from '../../../util/number';
+import { renderFromWei, weiToFiat, hexToBN, toBN, isBN, renderToGwei, balanceToFiat } from '../../../util/number';
 import Identicon from '../Identicon';
-import { getActionKey, decodeTransferData, getTicker, isCollectibleAddress } from '../../../util/transactions';
+import { getActionKey, decodeTransferData, getTicker } from '../../../util/transactions';
 import TransactionDetails from './TransactionDetails';
 import { renderFullAddress, safeToChecksumAddress } from '../../../util/address';
 import FadeIn from 'react-native-fade-in-image';
@@ -32,6 +19,7 @@ import StyledButton from '../StyledButton';
 import Networks from '../../../util/networks';
 import Device from '../../../util/Device';
 import Modal from 'react-native-modal';
+import { decodePaymentChannelTx, decodeTransferTx } from './utils';
 
 const {
 	CONNEXT: { CONTRACTS }
@@ -204,6 +192,7 @@ class TransactionElement extends PureComponent {
 		/**
 		 * Object containing token exchange rates in the format address => exchangeRate
 		 */
+		// eslint-disable-next-line react/no-unused-prop-types
 		contractExchangeRates: PropTypes.object,
 		/**
 		 * ETH to current currency conversion rate
@@ -228,6 +217,7 @@ class TransactionElement extends PureComponent {
 		/**
 		 * An array that represents the user tokens
 		 */
+		// eslint-disable-next-line react/no-unused-prop-types
 		tokens: PropTypes.object,
 		/**
 		 * An array that represents the user collectible contracts
@@ -287,11 +277,11 @@ class TransactionElement extends PureComponent {
 		const actionKey = tx.actionKey || (await getActionKey(tx, selectedAddress, ticker, paymentChannelTransaction));
 		let transactionElement, transactionDetails;
 		if (paymentChannelTransaction) {
-			[transactionElement, transactionDetails] = this.decodePaymentChannelTx();
+			[transactionElement, transactionDetails] = decodePaymentChannelTx({ ...this.props, actionKey });
 		} else {
 			switch (actionKey) {
 				case strings('transactions.sent_tokens'):
-					[transactionElement, transactionDetails] = await this.decodeTransferTx(actionKey);
+					[transactionElement, transactionDetails] = await decodeTransferTx({ ...this.props, actionKey });
 					break;
 				case strings('transactions.sent_collectible'):
 					[transactionElement, transactionDetails] = this.decodeTransferFromTx(actionKey);
@@ -646,146 +636,6 @@ class TransactionElement extends PureComponent {
 			paymentChannelTransaction: true
 		};
 
-		return [transactionElement, transactionDetails];
-	};
-
-	getTokenTransfer = totalGas => {
-		const {
-			tx: {
-				transaction: { to, data }
-			},
-			conversionRate,
-			currentCurrency,
-			tokens,
-			contractExchangeRates
-		} = this.props;
-
-		const { actionKey } = this.state;
-		const [, encodedAmount] = decodeTransferData('transfer', data);
-
-		const amount = toBN(encodedAmount);
-
-		const userHasToken = safeToChecksumAddress(to) in tokens;
-		const token = userHasToken ? tokens[safeToChecksumAddress(to)] : null;
-		const renderActionKey = token ? `${strings('transactions.sent')} ${token.symbol}` : actionKey;
-		const renderTokenAmount = token
-			? `${renderFromTokenMinimalUnit(amount, token.decimals)} ${token.symbol}`
-			: undefined;
-		const exchangeRate = token ? contractExchangeRates[token.address] : undefined;
-		let renderTokenFiatAmount, renderTokenFiatNumber;
-		if (exchangeRate) {
-			renderTokenFiatAmount = balanceToFiat(
-				fromTokenMinimalUnit(amount, token.decimals) || 0,
-				conversionRate,
-				exchangeRate,
-				currentCurrency
-			);
-			renderTokenFiatNumber = balanceToFiatNumber(
-				fromTokenMinimalUnit(amount, token.decimals) || 0,
-				conversionRate,
-				exchangeRate
-			);
-		}
-
-		const renderToken = token
-			? `${renderFromTokenMinimalUnit(amount, token.decimals)} ${token.symbol}`
-			: strings('transaction.value_not_available');
-		const totalFiatNumber = renderTokenFiatNumber
-			? weiToFiatNumber(totalGas, conversionRate) + renderTokenFiatNumber
-			: undefined;
-
-		const ticker = getTicker(this.props.ticker);
-
-		const transactionDetails = {
-			renderTotalGas: `${renderFromWei(totalGas)} ${ticker}`,
-			renderTotalGasFiat: weiToFiat(totalGas, conversionRate, currentCurrency),
-			renderValue: renderToken,
-			renderValueFiat: renderTokenFiatAmount ? `${renderTokenFiatAmount}` : undefined,
-			renderTotalValue: `${renderToken} ${strings('unit.divisor')} ${renderFromWei(totalGas)} ${ticker}`,
-			renderTotalValueFiat: totalFiatNumber ? `${addCurrencySymbol(totalFiatNumber, currentCurrency)}` : undefined
-		};
-
-		const transactionElement = {
-			actionKey: renderActionKey,
-			value: !renderTokenAmount ? strings('transaction.value_not_available') : renderTokenAmount,
-			fiatValue: `- ${renderTokenFiatAmount}`
-		};
-
-		return [transactionElement, transactionDetails];
-	};
-
-	getCollectibleTransfer = totalGas => {
-		const {
-			tx: {
-				transaction: { to }
-			},
-			collectibleContracts
-		} = this.props;
-		const { encodedAmount } = this.state;
-		let actionKey;
-		const tokenId = encodedAmount;
-		const collectible = collectibleContracts.find(
-			collectible => collectible.address.toLowerCase() === to.toLowerCase()
-		);
-		if (collectible) {
-			actionKey = `${strings('transactions.sent')} ${collectible.name}`;
-		} else {
-			actionKey = strings('transactions.sent_collectible');
-		}
-
-		const renderCollectible = collectible
-			? `${strings('unit.token_id')} ${tokenId} ${collectible.symbol}`
-			: `${strings('unit.token_id')} ${tokenId}`;
-
-		const transactionDetails = {
-			renderValue: renderCollectible,
-			renderTotalValue: `${renderCollectible} ${strings('unit.divisor')} ${renderFromWei(totalGas)} ${strings(
-				'unit.eth'
-			)}`,
-			renderTotalValueFiat: undefined
-		};
-
-		const transactionElement = {
-			actionKey,
-			value: `${strings('unit.token_id')}${tokenId}`,
-			fiatValue: collectible ? collectible.symbol : undefined
-		};
-
-		return [transactionElement, transactionDetails];
-	};
-
-	decodeTransferTx = async () => {
-		const {
-			tx: {
-				transaction: { from, gas, gasPrice, data, to },
-				transactionHash
-			}
-		} = this.props;
-
-		const decodedData = decodeTransferData('transfer', data);
-		const addressTo = decodedData[0];
-		const isCollectible = await isCollectibleAddress(to, decodedData[1]);
-
-		const gasBN = hexToBN(gas);
-		const gasPriceBN = hexToBN(gasPrice);
-		const totalGas = isBN(gasBN) && isBN(gasPriceBN) ? gasBN.mul(gasPriceBN) : toBN('0x0');
-		const renderGas = parseInt(gas, 16).toString();
-		const renderGasPrice = renderToGwei(gasPrice);
-
-		let [transactionElement, transactionDetails] = isCollectible
-			? this.getCollectibleTransfer(totalGas)
-			: this.getTokenTransfer(totalGas);
-		transactionElement = { ...transactionElement, renderTo: addressTo };
-		transactionDetails = {
-			...transactionDetails,
-			...{
-				renderFrom: renderFullAddress(from),
-				renderTo: renderFullAddress(addressTo),
-				transactionHash,
-				renderGas,
-				renderGasPrice
-			}
-		};
 		return [transactionElement, transactionDetails];
 	};
